Fix "cannot convert string to float64" log flood
When clients send numeric data in string format -- sometimes you just can't
control them -- the conversion logic apparently could not handle it correctly.
This led to a flurry of log output and, potentially worse, would just store
zero values in the database.

This patch checks if the input is of string type and passes it to
strconv.ParseFloat(). Can still go wrong, but then we haven't really lost
anything.